Add flag to track whether fast path is temporary
